accounts/phoenix-dev/us-west-2/route53/requirements.txt
aws-cdk-lib==2.22.0
constructs>=10.0.0,<11.0.0
